882 Add arrows (#381)
* Add arrows

* Move stlyes to a file

* Change file extension
