+++ src/shared/components/Graph/index.tsx
@@ -3,6 +3,8 @@ import * as cytoscape from 'cytoscape';
 import { Alert, Button } from 'antd';
 import * as cola from 'cytoscape-cola';
 
+import style from './style';
+
 import './GraphComponent.less';
 
 const LAYOUTS: {
@@ -110,80 +112,6 @@ const Graph: React.FunctionComponent<{
   const handleLayoutClick = (type: string) => () => {
     setLayoutType(type);
   };
-
-  const style = [
-    {
-      selector: 'node',
-      style: {
-        height: 16,
-        width: 16,
-      },
-    },
-    {
-      selector: '.-expandable',
-      style: {
-        height: 24,
-        width: 24,
-      },
-    },
-    {
-      selector: 'edge',
-      style: {
-        width: 2,
-        'target-arrow-shape': 'triangle',
-        'line-color': '#8a8b8b',
-      },
-    },
-    {
-      selector: 'node[label]',
-      style: {
-        'text-outline-color': 'white',
-        'text-outline-width': 2,
-        label: 'data(label)',
-      },
-    },
-    {
-      selector: 'edge[label]',
-      style: {
-        label: 'data(label)',
-        'text-outline-color': 'white',
-        'text-outline-width': 2,
-        'edge-text-rotation': 'autorotate',
-      },
-    },
-    {
-      selector: '.blank-node',
-      style: {
-        width: 12,
-        height: 12,
-        'background-color': 'white',
-        'border-color': '#faad14',
-        'border-width': 2,
-      },
-    },
-    {
-      selector: '.-external',
-      style: {
-        'background-color': 'white',
-        'border-color': '#00adee',
-        'border-width': 2,
-      },
-    },
-    {
-      selector: '.-internal',
-      style: {
-        'background-color': '#ff6666',
-      },
-    },
-    {
-      selector: '.-internal.-expanded',
-      style: {
-        'background-color': 'white',
-        'border-color': '#ff6666',
-        'border-width': 2,
-      },
-    },
-  ];
 
   const replaceElements = (elements: cytoscape.ElementDefinition[]) => {
     if (graph.current) {
